chore: Update airdrop/deploy.sh and hello-utility/deploy.sh scripts
The commit updates the airdrop/deploy.sh and hello-utility/deploy.sh scripts. In the airdrop/deploy.sh script, the ACCOUNT variable is changed from "${CONTRACT_ACCOUNT_ID:-unc}" to "${CONTRACT_ACCOUNT_ID:-drop}". This change ensures that the default value for the ACCOUNT variable is "drop" instead of "unc". In the hello-utility/deploy.sh script, there are no changes to the code, but the file is included in this commit to reflect the modifications made to both scripts.

These changes improve the consistency and clarity of the deployment process by updating the default value for the ACCOUNT variable in the airdrop/deploy.sh script and including the hello-utility/deploy.sh script in the commit.